requirements.txt
Click==7.0
Flask==1.1.1
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.17.2
Pillow==6.2.0
PyYAML==5.1.2
six==1.12.0
torch==1.2.0
torchvision==0.4.0
Werkzeug==0.16.0
